+++ management/db/migrate/20190707091938_create_order_books.rb
@@ -1,9 +1,9 @@
 class CreateOrderBooks < ActiveRecord::Migration[6.0]
   def change
     create_table :order_books, comment: '下单记录 包括ask and bid' do |t|
-      t.integer :user_id, null: false, comment: '用户'
+      t.bigint :user_id, null: false, comment: '用户'
       t.string :symbol, null: false, comment: '简称 eg BTC_USD'
-      t.integer :fund_id, null: false, comment: '商品'
+      t.bigint :fund_id, null: false, comment: '商品'
       t.integer :status, null: false, default: 0, comment: '状态'
       t.string :order_type, null: false, comment: '订单类型 市价单market 限价单limit'
       t.string :side, null: false, comment: 'sell or buy'
